requirements.txt
alembic==1.7.5
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
attrs==21.4.0
cffi==1.15.0
click==8.0.3
flake8==4.0.1
Flask==2.0.2
Flask-JWT-Extended==4.3.1
greenlet==1.1.2
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.0.1
Jinja2==3.0.3
Mako==1.1.6
MarkupSafe==2.0.1
marshmallow==3.14.1
mccabe==0.6.1
packaging==21.3
passlib==1.7.4
pluggy==1.0.0
psycopg2==2.9.3
py==1.11.0
pycodestyle==2.8.0
pycparser==2.21
pyflakes==2.4.0
PyJWT==2.3.0
pyparsing==3.0.7
pytest==6.2.5
python-dotenv==0.19.2
six==1.16.0
sqla-softdelete==1.2
SQLAlchemy==1.4.31
SQLAlchemy-Utils==0.38.2
toml==0.10.2
Werkzeug==2.0.2
